+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## lifebit-ai/cloudos-cli: changelog
 
+### 1.3.0 - 2022-11-07
+- All Cromwell functionality works now with personal API key. The
+`--cromwell-token` argument is maintained for backwards compatibility, but can
+be completely substituted by `--apikey`.
+- Changes `--wdl-importsfile` parameter to be optional even when running a
+WDL pipeline as `importsFiles` are not always present in WDL pipelines.
+- Fixes some incomplete error messages.
+
 ### 1.2.1 - 2022-11-03
 - Modifies default `--cost-limit` from infinite (`-1`) to `30.0`. This will prevent
 wasting resources without a purpose of running a pipeline.